payment-service/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
certifi
